Added NBT-Values and did some code refactoring
Removed deserialize-method from Serializable interface. There is a note, which tells the user to add a static deserialize-method.
Added class Wrapper for storing the wrapped object in the wrapper object.
Fixed a bug affecting the NBTReader, which stopped the NBT-Reader from indexing any NBTTagEntry, which was lower then two NBTTagCompounds.
Added a BukkitRunnable to the PacketReader.
Added a test-item, which can be used to test new features affecting items.