Added durability for items.
Added durability for items.
Added new method for the NBT-Reader to scan for sub-tags more effective using recursion.
Added BlockBreakListener to remove durability from an item, if you try to break a block using a tool or to cancel the block breaking if the tool has no durability.
Added durability-operations to the EntityDamageByEntityListener.
Added method for updating the items lore, name, durability values and damage-stats.
Added a durability of 250 to the crop-armor.