Added option to set a custom operation, if a block gets clicked.
Added option to run custom code if a block gets clicked.
Added method in Reflection to scan a package for all classes, that extend a certain other class.
Added a map of reference items, that stores all items only one time. This map is used for getting an items ClickAction after it got serialized or built.