Added resizable enderchests.
Added me.trqhxrd:menus:1.1-SNAPSHOT, which is a better version of the GrapesInventory. In the future, that will be the only menu library we use. The migration probably will be done in the next commit.
Added attribute Savable, which, when implemented, adds method for saving an object.
Renamed package "me.trqhxrd.grapesrpg.game.inventories" to "me.trqhxrd.grapesrpg.game.inventory".
Added some package-infos