Added Artifacts and updated NBT-API to use Objects instead of NBTValues because of serialisation problems.
Added Run-Config for Breakpoint-usage.
Moved all the GrapesEvents regarding players to their own package.
Added Artifacts, which exist only once per server.
Removed NBTValue because of serialisation problems. Objects will be used instead.
Added ArtifactConfig, which stores all the different artifacts and is also in charge of loading them.
Added a bunch of listeners, which make sure the states of all of the artifacts are updated correctly.
Also there was a project-wide code cleanup